Added a Keras Text Generation Example
The Keras text generation example is enclosed in triple quotes as a multi-line comment near the bottom of textencodefunc.py. There is a bug that is ruining the neural network's outputs. The strangely similar outputs do not appear to be coming from the neural network, but rather from some stage of encoding. I would appreciate it if anyone can find and squash this bug :)